requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
WTForms==3.1.1
supabase==2.3.0
pydantic==2.10.5
python-dotenv==1.0.0
email-validator==2.1.0
cryptography>=41.0.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
black==23.11.0
ruff==0.1.6
gunicorn==21.2.0
openai==1.3.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
beautifulsoup4==4.12.2
requests==2.31.0
